Triple oxygen isotope values

d17O_c() calculates the equilibrium d18O, d17O, and D17O values of a calcite grown at a given temperature.

d17O_c(temp, d18O_H2O_VSMOW, eq18 = "Daeron19", lambda = 0.528)

Arguments

temp
Calcite growth temperature (°C).

d18O_H2O_VSMOW
Water d18O value expressed on the VSMOW scale (‰).

eq18
Equation used to calculate the 18O/16O fractionation factor between calcite and water. Options are like those for calcite in a18_c_H2O() with "Daeron19" being here the default.

lambda
Triple oxygen isotope reference slope. Default 0.528.

Value

Returns a data frame:

d18O value of the carbonate expressed on the VSMOW scale (‰).

d18O value of the carbonate expressed on the VSMOW scale (‰).

D17O value of the carbonate expressed on the VSMOW scale (‰).

Details

$$\theta_{A/B} = \frac{\alpha^{17}_{A/B}}{\alpha^{18}_{A/B}}$$

$$ \delta'^{17}O_{H2O,VSMOW} = \beta \times \delta'^{18}O_{H2O,VSMOW} + \gamma \textrm{ , where } \beta=0.528 \textrm{ and } \gamma = 0 $$

$$\Delta^{17}O_{CaCO3,VSMOW} = \delta'^{17}O_{CaCO3,VSMOW} - \lambda \times \delta'^{18}O_{CaCO3,VSMOW} $$

References

Guo, W., & Zhou, C. (2019). Triple oxygen isotope fractionation in the DIC-H2O-CO2 system: A numerical framework and its implications. Geochimica et Cosmochimica Acta, 246, 541-564. doi: 10.1016/j.gca.2018.11.018

See also

Examples

d17O_c(temp = 10, d18O_H2O_VSMOW = -1) # Returns the data frame (length = 3)
#>     d18O_c   d17O_c      D17O_c
#> 1 32.43637 16.91235 -0.08357564
prime(d17O_c(temp = 10, d18O_H2O_VSMOW = -1)[, 2]) # Returns the d'17O value
#> [1] 16.77093
d17O_c(temp = 10, d18O_H2O_VSMOW = -1)[, 3] # Returns the D17O value
#> [1] -0.08357564