Triple oxygen isotope values of quartz

d17O_qz() calculates the equilibrium d18O, d17O, and D17O values of quartz grown at a given temperature.

d17O_qz(temp, d18O_H2O_VSMOW, D17O_H2O = 0, lambda = 0.528)

Arguments

temp

Quartz growth temperature (°C).

d18O_H2O_VSMOW

Water d18O value expressed on the VSMOW scale (‰).

D17O_H2O

D17O value of ambient water calculated using a lambda of 0.528. Default 0.

lambda

Triple oxygen isotope reference slope. Default 0.528.

Value

Returns a data frame:

d18O value of the quartz expressed on the VSMOW scale (‰).

d18O value of the quartz expressed on the VSMOW scale (‰).

D17O value of the quartz expressed on the VSMOW scale (‰).

Details

$$\theta_{A/B} = \frac{\alpha^{17}_{A/B}}{\alpha^{18}_{A/B}}$$

$$ \delta'^{17}O_{H2O,VSMOW} = \beta \times \delta'^{18}O_{H2O,VSMOW} + \gamma \textrm{ , where } \beta=0.528 \textrm{ and } \gamma = 0 $$

$$\Delta^{17}O_{SiO2,VSMOW} = \delta'^{17}O_{SiO2,VSMOW} - \lambda \times \delta'^{18}O_{SiO2,VSMOW} $$

References

Sharp, Z.D., Gibbons, J.A., Maltsev, O., Atudorei, V., Pack, A., Sengupta, S., Shock, E.L. and Knauth, L.P. (2016). A calibration of the triple oxygen isotope fractionation in the SiO2-H2O system and applications to natural samples. Geochimica et Cosmochimica Acta, 186, 105-119. doi:10.1016/j.gca.2016.04.047

Examples

d17O_qz(temp = 10, d18O_H2O_VSMOW = 0) # Returns the data frame (length = 3)
#>    d18O_qz  d17O_qz   D17O_qz
#> 1 41.87607 21.72735 -0.165472
d17O_qz(temp = 10, d18O_H2O_VSMOW = 0)[, 3] # Returns the D17O value
#> [1] -0.165472