Water d18O value

d18O_H2O() calculates the d18O value of the ambient water from the d18O value of a carbonate and its growth temperature.

d18O_H2O(temp, d18O_c_VSMOW, min, eq)

Arguments

temp

Carbonate growth temperature (°C).

d18O_c_VSMOW

Carbonate d18O value expressed on the VSMOW scale (‰).

min

Mineralogy. Options are "calcite", "aragonite", and "dolomite".

eq

Equation used to calculate the equilibrium 18O/16O oxygen isotope fractionation factor between carbonate and water. Options depend on mineralogy and listed in a18_c_H2O().

Value

Returns the water d18O value expressed on the VSMOW scale (‰).

Note

Use to_VSMOW() and to_VPDB() to convert between the VSMOW and VPDB scales.

References

References are listed in the description of a18_c_H2O().

See also

Examples

d18O_H2O(temp = 33.7, d18O_c_VSMOW = 14.58,
         min = "calcite", eq = "Coplen07")
#> [1] -13.53802
d18O_H2O(temp = 25, d18O_c_VSMOW = to_VSMOW(-7.47),
         min = "aragonite", eq = "GK86")
#> [1] -6.527318