Equilibrium carbonate d18O value

d18O_c() calculates the equilibrium d18O value of a carbonate grown at a given temperature.

Usage

d18O_c(temp, d18O_H2O_VSMOW, min, eq)

Arguments

temp

Carbonate growth temperature (°C).

d18O_H2O_VSMOW

Water d18O value expressed on the VSMOW scale (‰).

min

Mineralogy. Options are as in a18_c_H2O().

eq

Equation used for the calculations. Options depend on mineralogy and are listed in a18_c_H2O().

Value

Returns the equilibrium carbonate d18O value expressed on the VSMOW scale (‰).

Note

Use to_VSMOW() and to_VPDB() to convert between the VSMOW and VPDB scales.

References

References are listed in the description of a18_c_H2O().

See also

Examples

d18O_c(33.7, -13.54, min = "calcite", eq = "Coplen07")
#> [1] 14.57797
to_VPDB(d18O_c(temp = 12, d18O_H2O_VSMOW = -6.94,
               min = "aragonite", eq = "GK86"))
#> [1] -5.212518