Regression confidence intervals

york_plot() calculates and optionally plots the confidence intervals of an (error-considering) linear regression.

york_plot(
  x,
  slope,
  slope_se,
  intercept,
  intercept_se,
  cl = 0.95,
  weights = -1,
  add = FALSE,
  col = "black"
)

Arguments

x

x values of the data points.

slope

regression slope.

slope_se

Standard error of the slope.

intercept

regression intercept.

intercept_se

Standard error of the intercept.

cl

Confidence level. Default: 0.95.

weights

Weights of the data points. If given, mean & SD of x are computed with the weights. Has to be same the length as x. Optional.

add

Add graphics to an already existing plot? Default: FALSE.

col

Graphical parameter. Optional.

Value

A list with regression parameters:

slope and its standard error

intercept and its standard error

weights of the points (normalized to 1)

residual standard error (sigma)

R2

p-value (two-tailed t-test).

Details

The algorithm is described in the appendix of Wacker et al. (2014).

Contributors

Julian Tödter

References

Wacker, U., Fiebig, J., Tödter, J., Schöne, B. R., Bahr, A., Friedrich, O., et al. (2014). Empirical calibration of the clumped isotope paleothermometer using calcites of various origins. Geochimica et Cosmochimica Acta, 141, 127-144. doi:10.1016/j.gca.2014.06.004

Examples

york_plot(
  x = c(1, 2, 3),
  slope = 1.06,
  slope_se = 1.60,
  intercept = -0.05,
  intercept_se = 0.34,
  cl = 0.98)
#>             x     y_low     y_up
#> 1    1.000000 -35.08687 37.10687
#> 2    1.002002 -35.04141 37.06565
#> 3    1.004004 -34.99598 37.02447
#> 4    1.006006 -34.95059 36.98332
#> 5    1.008008 -34.90523 36.94220
#> 6    1.010010 -34.85990 36.90112
#> 7    1.012012 -34.81461 36.86008
#> 8    1.014014 -34.76936 36.81907
#> 9    1.016016 -34.72414 36.77810
#> 10   1.018018 -34.67896 36.73716
#> 11   1.020020 -34.63381 36.69626
#> 12   1.022022 -34.58870 36.65539
#> 13   1.024024 -34.54363 36.61456
#> 14   1.026026 -34.49859 36.57377
#> 15   1.028028 -34.45359 36.53301
#> 16   1.030030 -34.40863 36.49229
#> 17   1.032032 -34.36370 36.45161
#> 18   1.034034 -34.31881 36.41096
#> 19   1.036036 -34.27395 36.37035
#> 20   1.038038 -34.22914 36.32978
#> 21   1.040040 -34.18435 36.28924
#> 22   1.042042 -34.13961 36.24874
#> 23   1.044044 -34.09491 36.20828
#> 24   1.046046 -34.05024 36.16786
#> 25   1.048048 -34.00561 36.12747
#> 26   1.050050 -33.96102 36.08712
#> 27   1.052052 -33.91646 36.04681
#> 28   1.054054 -33.87195 36.00654
#> 29   1.056056 -33.82747 35.96631
#> 30   1.058058 -33.78303 35.92611
#> 31   1.060060 -33.73863 35.88596
#> 32   1.062062 -33.69427 35.84584
#> 33   1.064064 -33.64994 35.80576
#> 34   1.066066 -33.60566 35.76572
#> 35   1.068068 -33.56141 35.72572
#> 36   1.070070 -33.51721 35.68576
#> 37   1.072072 -33.47304 35.64583
#> 38   1.074074 -33.42891 35.60595
#> 39   1.076076 -33.38483 35.56611
#> 40   1.078078 -33.34078 35.52631
#> 41   1.080080 -33.29677 35.48654
#> 42   1.082082 -33.25281 35.44682
#> 43   1.084084 -33.20888 35.40714
#> 44   1.086086 -33.16499 35.36750
#> 45   1.088088 -33.12115 35.32789
#> 46   1.090090 -33.07734 35.28833
#> 47   1.092092 -33.03358 35.24881
#> 48   1.094094 -32.98985 35.20933
#> 49   1.096096 -32.94617 35.16990
#> 50   1.098098 -32.90253 35.13050
#> 51   1.100100 -32.85893 35.09115
#> 52   1.102102 -32.81538 35.05183
#> 53   1.104104 -32.77186 35.01256
#> 54   1.106106 -32.72839 34.97333
#> 55   1.108108 -32.68495 34.93414
#> 56   1.110110 -32.64157 34.89500
#> 57   1.112112 -32.59822 34.85590
#> 58   1.114114 -32.55491 34.81684
#> 59   1.116116 -32.51165 34.77782
#> 60   1.118118 -32.46843 34.73884
#> 61   1.120120 -32.42526 34.69991
#> 62   1.122122 -32.38213 34.66102
#> 63   1.124124 -32.33904 34.62218
#> 64   1.126126 -32.29599 34.58338
#> 65   1.128128 -32.25299 34.54462
#> 66   1.130130 -32.21003 34.50591
#> 67   1.132132 -32.16712 34.46724
#> 68   1.134134 -32.12425 34.42861
#> 69   1.136136 -32.08142 34.39003
#> 70   1.138138 -32.03864 34.35149
#> 71   1.140140 -31.99590 34.31300
#> 72   1.142142 -31.95321 34.27455
#> 73   1.144144 -31.91056 34.23615
#> 74   1.146146 -31.86796 34.19779
#> 75   1.148148 -31.82541 34.15948
#> 76   1.150150 -31.78290 34.12121
#> 77   1.152152 -31.74043 34.08299
#> 78   1.154154 -31.69801 34.04482
#> 79   1.156156 -31.65564 34.00669
#> 80   1.158158 -31.61331 33.96861
#> 81   1.160160 -31.57103 33.93057
#> 82   1.162162 -31.52880 33.89258
#> 83   1.164164 -31.48661 33.85464
#> 84   1.166166 -31.44447 33.81674
#> 85   1.168168 -31.40237 33.77889
#> 86   1.170170 -31.36033 33.74109
#> 87   1.172172 -31.31833 33.70333
#> 88   1.174174 -31.27637 33.66562
#> 89   1.176176 -31.23447 33.62796
#> 90   1.178178 -31.19261 33.59035
#> 91   1.180180 -31.15081 33.55279
#> 92   1.182182 -31.10905 33.51527
#> 93   1.184184 -31.06733 33.47780
#> 94   1.186186 -31.02567 33.44039
#> 95   1.188188 -30.98406 33.40302
#> 96   1.190190 -30.94249 33.36570
#> 97   1.192192 -30.90098 33.32842
#> 98   1.194194 -30.85951 33.29120
#> 99   1.196196 -30.81809 33.25403
#> 100  1.198198 -30.77672 33.21690
#> 101  1.200200 -30.73541 33.17983
#> 102  1.202202 -30.69414 33.14281
#> 103  1.204204 -30.65292 33.10583
#> 104  1.206206 -30.61175 33.06891
#> 105  1.208208 -30.57064 33.03204
#> 106  1.210210 -30.52957 32.99522
#> 107  1.212212 -30.48856 32.95845
#> 108  1.214214 -30.44759 32.92173
#> 109  1.216216 -30.40668 32.88506
#> 110  1.218218 -30.36582 32.84844
#> 111  1.220220 -30.32501 32.81188
#> 112  1.222222 -30.28425 32.77536
#> 113  1.224224 -30.24355 32.73890
#> 114  1.226226 -30.20289 32.70249
#> 115  1.228228 -30.16229 32.66614
#> 116  1.230230 -30.12175 32.62984
#> 117  1.232232 -30.08125 32.59358
#> 118  1.234234 -30.04081 32.55739
#> 119  1.236236 -30.00042 32.52124
#> 120  1.238238 -29.96009 32.48515
#> 121  1.240240 -29.91981 32.44912
#> 122  1.242242 -29.87958 32.41313
#> 123  1.244244 -29.83940 32.37720
#> 124  1.246246 -29.79929 32.34133
#> 125  1.248248 -29.75922 32.30551
#> 126  1.250250 -29.71921 32.26974
#> 127  1.252252 -29.67926 32.23403
#> 128  1.254254 -29.63936 32.19837
#> 129  1.256256 -29.59951 32.16277
#> 130  1.258258 -29.55972 32.12723
#> 131  1.260260 -29.51999 32.09174
#> 132  1.262262 -29.48031 32.05631
#> 133  1.264264 -29.44069 32.02093
#> 134  1.266266 -29.40112 31.98561
#> 135  1.268268 -29.36161 31.95034
#> 136  1.270270 -29.32216 31.91513
#> 137  1.272272 -29.28276 31.87998
#> 138  1.274274 -29.24343 31.84489
#> 139  1.276276 -29.20415 31.80985
#> 140  1.278278 -29.16492 31.77487
#> 141  1.280280 -29.12576 31.73995
#> 142  1.282282 -29.08665 31.70509
#> 143  1.284284 -29.04760 31.67028
#> 144  1.286286 -29.00861 31.63553
#> 145  1.288288 -28.96967 31.60084
#> 146  1.290290 -28.93080 31.56622
#> 147  1.292292 -28.89198 31.53164
#> 148  1.294294 -28.85323 31.49713
#> 149  1.296296 -28.81453 31.46268
#> 150  1.298298 -28.77590 31.42829
#> 151  1.300300 -28.73732 31.39396
#> 152  1.302302 -28.69880 31.35968
#> 153  1.304304 -28.66035 31.32547
#> 154  1.306306 -28.62195 31.29132
#> 155  1.308308 -28.58361 31.25723
#> 156  1.310310 -28.54534 31.22320
#> 157  1.312312 -28.50713 31.18923
#> 158  1.314314 -28.46898 31.15532
#> 159  1.316316 -28.43089 31.12148
#> 160  1.318318 -28.39286 31.08769
#> 161  1.320320 -28.35489 31.05397
#> 162  1.322322 -28.31699 31.02031
#> 163  1.324324 -28.27915 30.98671
#> 164  1.326326 -28.24137 30.95318
#> 165  1.328328 -28.20365 30.91971
#> 166  1.330330 -28.16600 30.88630
#> 167  1.332332 -28.12841 30.85296
#> 168  1.334334 -28.09089 30.81968
#> 169  1.336336 -28.05343 30.78646
#> 170  1.338338 -28.01603 30.75331
#> 171  1.340340 -27.97870 30.72022
#> 172  1.342342 -27.94143 30.68719
#> 173  1.344344 -27.90422 30.65423
#> 174  1.346346 -27.86709 30.62134
#> 175  1.348348 -27.83001 30.58851
#> 176  1.350350 -27.79301 30.55575
#> 177  1.352352 -27.75606 30.52305
#> 178  1.354354 -27.71919 30.49042
#> 179  1.356356 -27.68238 30.45786
#> 180  1.358358 -27.64564 30.42536
#> 181  1.360360 -27.60896 30.39292
#> 182  1.362362 -27.57235 30.36056
#> 183  1.364364 -27.53581 30.32826
#> 184  1.366366 -27.49933 30.29603
#> 185  1.368368 -27.46293 30.26387
#> 186  1.370370 -27.42659 30.23177
#> 187  1.372372 -27.39032 30.19975
#> 188  1.374374 -27.35412 30.16779
#> 189  1.376376 -27.31798 30.13590
#> 190  1.378378 -27.28192 30.10408
#> 191  1.380380 -27.24592 30.07233
#> 192  1.382382 -27.21000 30.04065
#> 193  1.384384 -27.17414 30.00903
#> 194  1.386386 -27.13835 29.97749
#> 195  1.388388 -27.10263 29.94602
#> 196  1.390390 -27.06699 29.91462
#> 197  1.392392 -27.03141 29.88328
#> 198  1.394394 -26.99591 29.85202
#> 199  1.396396 -26.96047 29.82083
#> 200  1.398398 -26.92511 29.78972
#> 201  1.400400 -26.88982 29.75867
#> 202  1.402402 -26.85460 29.72769
#> 203  1.404404 -26.81945 29.69679
#> 204  1.406406 -26.78438 29.66596
#> 205  1.408408 -26.74938 29.63520
#> 206  1.410410 -26.71445 29.60452
#> 207  1.412412 -26.67959 29.57391
#> 208  1.414414 -26.64481 29.54337
#> 209  1.416416 -26.61010 29.51291
#> 210  1.418418 -26.57547 29.48252
#> 211  1.420420 -26.54091 29.45220
#> 212  1.422422 -26.50642 29.42196
#> 213  1.424424 -26.47201 29.39179
#> 214  1.426426 -26.43767 29.36170
#> 215  1.428428 -26.40341 29.33168
#> 216  1.430430 -26.36923 29.30174
#> 217  1.432432 -26.33512 29.27187
#> 218  1.434434 -26.30108 29.24208
#> 219  1.436436 -26.26712 29.21237
#> 220  1.438438 -26.23324 29.18273
#> 221  1.440440 -26.19944 29.15317
#> 222  1.442442 -26.16571 29.12369
#> 223  1.444444 -26.13206 29.09428
#> 224  1.446446 -26.09849 29.06495
#> 225  1.448448 -26.06499 29.03570
#> 226  1.450450 -26.03157 29.00653
#> 227  1.452452 -25.99824 28.97743
#> 228  1.454454 -25.96498 28.94842
#> 229  1.456456 -25.93179 28.91948
#> 230  1.458458 -25.89869 28.89062
#> 231  1.460460 -25.86567 28.86185
#> 232  1.462462 -25.83273 28.83315
#> 233  1.464464 -25.79986 28.80453
#> 234  1.466466 -25.76708 28.77599
#> 235  1.468468 -25.73438 28.74753
#> 236  1.470470 -25.70176 28.71916
#> 237  1.472472 -25.66922 28.69086
#> 238  1.474474 -25.63676 28.66265
#> 239  1.476476 -25.60438 28.63451
#> 240  1.478478 -25.57209 28.60646
#> 241  1.480480 -25.53987 28.57849
#> 242  1.482482 -25.50774 28.55061
#> 243  1.484484 -25.47570 28.52280
#> 244  1.486486 -25.44373 28.49508
#> 245  1.488488 -25.41185 28.46745
#> 246  1.490490 -25.38005 28.43989
#> 247  1.492492 -25.34834 28.41242
#> 248  1.494494 -25.31671 28.38503
#> 249  1.496496 -25.28516 28.35773
#> 250  1.498498 -25.25370 28.33052
#> 251  1.500501 -25.22232 28.30338
#> 252  1.502503 -25.19103 28.27634
#> 253  1.504505 -25.15983 28.24938
#> 254  1.506507 -25.12871 28.22250
#> 255  1.508509 -25.09767 28.19571
#> 256  1.510511 -25.06672 28.16901
#> 257  1.512513 -25.03586 28.14239
#> 258  1.514515 -25.00509 28.11586
#> 259  1.516517 -24.97440 28.08942
#> 260  1.518519 -24.94380 28.06306
#> 261  1.520521 -24.91329 28.03679
#> 262  1.522523 -24.88287 28.01061
#> 263  1.524525 -24.85253 27.98452
#> 264  1.526527 -24.82228 27.95852
#> 265  1.528529 -24.79213 27.93261
#> 266  1.530531 -24.76206 27.90678
#> 267  1.532533 -24.73208 27.88105
#> 268  1.534535 -24.70219 27.85540
#> 269  1.536537 -24.67239 27.82985
#> 270  1.538539 -24.64268 27.80438
#> 271  1.540541 -24.61306 27.77900
#> 272  1.542543 -24.58353 27.75372
#> 273  1.544545 -24.55409 27.72853
#> 274  1.546547 -24.52475 27.70343
#> 275  1.548549 -24.49549 27.67842
#> 276  1.550551 -24.46633 27.65350
#> 277  1.552553 -24.43726 27.62867
#> 278  1.554555 -24.40829 27.60394
#> 279  1.556557 -24.37940 27.57930
#> 280  1.558559 -24.35061 27.55475
#> 281  1.560561 -24.32191 27.53030
#> 282  1.562563 -24.29331 27.50594
#> 283  1.564565 -24.26480 27.48168
#> 284  1.566567 -24.23638 27.45750
#> 285  1.568569 -24.20806 27.43343
#> 286  1.570571 -24.17984 27.40945
#> 287  1.572573 -24.15170 27.38556
#> 288  1.574575 -24.12367 27.36177
#> 289  1.576577 -24.09573 27.33807
#> 290  1.578579 -24.06789 27.31447
#> 291  1.580581 -24.04014 27.29097
#> 292  1.582583 -24.01249 27.26756
#> 293  1.584585 -23.98493 27.24425
#> 294  1.586587 -23.95748 27.22104
#> 295  1.588589 -23.93012 27.19792
#> 296  1.590591 -23.90285 27.17491
#> 297  1.592593 -23.87569 27.15199
#> 298  1.594595 -23.84863 27.12917
#> 299  1.596597 -23.82166 27.10644
#> 300  1.598599 -23.79479 27.08382
#> 301  1.600601 -23.76802 27.06130
#> 302  1.602603 -23.74135 27.03887
#> 303  1.604605 -23.71478 27.01655
#> 304  1.606607 -23.68831 26.99432
#> 305  1.608609 -23.66195 26.97220
#> 306  1.610611 -23.63568 26.95017
#> 307  1.612613 -23.60951 26.92825
#> 308  1.614615 -23.58344 26.90643
#> 309  1.616617 -23.55748 26.88471
#> 310  1.618619 -23.53162 26.86309
#> 311  1.620621 -23.50586 26.84157
#> 312  1.622623 -23.48020 26.82016
#> 313  1.624625 -23.45464 26.79885
#> 314  1.626627 -23.42919 26.77764
#> 315  1.628629 -23.40384 26.75653
#> 316  1.630631 -23.37859 26.73553
#> 317  1.632633 -23.35345 26.71463
#> 318  1.634635 -23.32841 26.69384
#> 319  1.636637 -23.30348 26.67315
#> 320  1.638639 -23.27865 26.65256
#> 321  1.640641 -23.25393 26.63208
#> 322  1.642643 -23.22931 26.61171
#> 323  1.644645 -23.20479 26.59144
#> 324  1.646647 -23.18039 26.57128
#> 325  1.648649 -23.15609 26.55122
#> 326  1.650651 -23.13189 26.53127
#> 327  1.652653 -23.10780 26.51143
#> 328  1.654655 -23.08382 26.49169
#> 329  1.656657 -23.05995 26.47206
#> 330  1.658659 -23.03618 26.45254
#> 331  1.660661 -23.01252 26.43312
#> 332  1.662663 -22.98897 26.41381
#> 333  1.664665 -22.96553 26.39462
#> 334  1.666667 -22.94219 26.37553
#> 335  1.668669 -22.91897 26.35655
#> 336  1.670671 -22.89585 26.33767
#> 337  1.672673 -22.87285 26.31891
#> 338  1.674675 -22.84995 26.30026
#> 339  1.676677 -22.82716 26.28172
#> 340  1.678679 -22.80449 26.26329
#> 341  1.680681 -22.78192 26.24496
#> 342  1.682683 -22.75947 26.22675
#> 343  1.684685 -22.73712 26.20865
#> 344  1.686687 -22.71489 26.19067
#> 345  1.688689 -22.69277 26.17279
#> 346  1.690691 -22.67076 26.15503
#> 347  1.692693 -22.64887 26.13737
#> 348  1.694695 -22.62708 26.11984
#> 349  1.696697 -22.60541 26.10241
#> 350  1.698699 -22.58385 26.08509
#> 351  1.700701 -22.56241 26.06789
#> 352  1.702703 -22.54108 26.05081
#> 353  1.704705 -22.51986 26.03384
#> 354  1.706707 -22.49876 26.01698
#> 355  1.708709 -22.47777 26.00023
#> 356  1.710711 -22.45690 25.98360
#> 357  1.712713 -22.43614 25.96709
#> 358  1.714715 -22.41549 25.95069
#> 359  1.716717 -22.39497 25.93440
#> 360  1.718719 -22.37455 25.91824
#> 361  1.720721 -22.35426 25.90218
#> 362  1.722723 -22.33408 25.88625
#> 363  1.724725 -22.31401 25.87043
#> 364  1.726727 -22.29407 25.85473
#> 365  1.728729 -22.27424 25.83914
#> 366  1.730731 -22.25452 25.82367
#> 367  1.732733 -22.23493 25.80832
#> 368  1.734735 -22.21545 25.79309
#> 369  1.736737 -22.19609 25.77797
#> 370  1.738739 -22.17685 25.76298
#> 371  1.740741 -22.15773 25.74810
#> 372  1.742743 -22.13873 25.73334
#> 373  1.744745 -22.11984 25.71870
#> 374  1.746747 -22.10108 25.70418
#> 375  1.748749 -22.08243 25.68978
#> 376  1.750751 -22.06391 25.67550
#> 377  1.752753 -22.04550 25.66134
#> 378  1.754755 -22.02721 25.64730
#> 379  1.756757 -22.00905 25.63337
#> 380  1.758759 -21.99101 25.61958
#> 381  1.760761 -21.97308 25.60590
#> 382  1.762763 -21.95528 25.59234
#> 383  1.764765 -21.93760 25.57890
#> 384  1.766767 -21.92004 25.56559
#> 385  1.768769 -21.90261 25.55240
#> 386  1.770771 -21.88530 25.53933
#> 387  1.772773 -21.86810 25.52638
#> 388  1.774775 -21.85104 25.51356
#> 389  1.776777 -21.83409 25.50086
#> 390  1.778779 -21.81727 25.48828
#> 391  1.780781 -21.80057 25.47582
#> 392  1.782783 -21.78399 25.46349
#> 393  1.784785 -21.76754 25.45129
#> 394  1.786787 -21.75121 25.43920
#> 395  1.788789 -21.73501 25.42724
#> 396  1.790791 -21.71893 25.41541
#> 397  1.792793 -21.70298 25.40370
#> 398  1.794795 -21.68715 25.39212
#> 399  1.796797 -21.67145 25.38066
#> 400  1.798799 -21.65587 25.36932
#> 401  1.800801 -21.64042 25.35811
#> 402  1.802803 -21.62509 25.34703
#> 403  1.804805 -21.60989 25.33608
#> 404  1.806807 -21.59481 25.32524
#> 405  1.808809 -21.57987 25.31454
#> 406  1.810811 -21.56505 25.30396
#> 407  1.812813 -21.55035 25.29351
#> 408  1.814815 -21.53578 25.28319
#> 409  1.816817 -21.52134 25.27300
#> 410  1.818819 -21.50703 25.26293
#> 411  1.820821 -21.49285 25.25299
#> 412  1.822823 -21.47879 25.24318
#> 413  1.824825 -21.46486 25.23349
#> 414  1.826827 -21.45106 25.22393
#> 415  1.828829 -21.43739 25.21451
#> 416  1.830831 -21.42385 25.20521
#> 417  1.832833 -21.41043 25.19604
#> 418  1.834835 -21.39715 25.18700
#> 419  1.836837 -21.38399 25.17809
#> 420  1.838839 -21.37096 25.16930
#> 421  1.840841 -21.35807 25.16065
#> 422  1.842843 -21.34530 25.15213
#> 423  1.844845 -21.33266 25.14373
#> 424  1.846847 -21.32016 25.13547
#> 425  1.848849 -21.30778 25.12734
#> 426  1.850851 -21.29553 25.11934
#> 427  1.852853 -21.28342 25.11147
#> 428  1.854855 -21.27143 25.10373
#> 429  1.856857 -21.25958 25.09612
#> 430  1.858859 -21.24786 25.08864
#> 431  1.860861 -21.23627 25.08129
#> 432  1.862863 -21.22481 25.07407
#> 433  1.864865 -21.21348 25.06699
#> 434  1.866867 -21.20228 25.06004
#> 435  1.868869 -21.19122 25.05322
#> 436  1.870871 -21.18028 25.04653
#> 437  1.872873 -21.16948 25.03997
#> 438  1.874875 -21.15882 25.03355
#> 439  1.876877 -21.14828 25.02726
#> 440  1.878879 -21.13788 25.02110
#> 441  1.880881 -21.12761 25.01508
#> 442  1.882883 -21.11747 25.00918
#> 443  1.884885 -21.10747 25.00343
#> 444  1.886887 -21.09760 24.99780
#> 445  1.888889 -21.08786 24.99231
#> 446  1.890891 -21.07826 24.98695
#> 447  1.892893 -21.06879 24.98172
#> 448  1.894895 -21.05945 24.97663
#> 449  1.896897 -21.05025 24.97167
#> 450  1.898899 -21.04118 24.96685
#> 451  1.900901 -21.03225 24.96216
#> 452  1.902903 -21.02345 24.95760
#> 453  1.904905 -21.01479 24.95318
#> 454  1.906907 -21.00625 24.94890
#> 455  1.908909 -20.99786 24.94475
#> 456  1.910911 -20.98960 24.94073
#> 457  1.912913 -20.98147 24.93685
#> 458  1.914915 -20.97348 24.93310
#> 459  1.916917 -20.96562 24.92949
#> 460  1.918919 -20.95790 24.92601
#> 461  1.920921 -20.95032 24.92267
#> 462  1.922923 -20.94287 24.91946
#> 463  1.924925 -20.93555 24.91639
#> 464  1.926927 -20.92837 24.91346
#> 465  1.928929 -20.92133 24.91066
#> 466  1.930931 -20.91442 24.90799
#> 467  1.932933 -20.90765 24.90546
#> 468  1.934935 -20.90101 24.90307
#> 469  1.936937 -20.89451 24.90081
#> 470  1.938939 -20.88814 24.89869
#> 471  1.940941 -20.88191 24.89671
#> 472  1.942943 -20.87582 24.89486
#> 473  1.944945 -20.86987 24.89315
#> 474  1.946947 -20.86405 24.89157
#> 475  1.948949 -20.85836 24.89013
#> 476  1.950951 -20.85282 24.88883
#> 477  1.952953 -20.84741 24.88767
#> 478  1.954955 -20.84213 24.88664
#> 479  1.956957 -20.83699 24.88574
#> 480  1.958959 -20.83199 24.88499
#> 481  1.960961 -20.82713 24.88437
#> 482  1.962963 -20.82240 24.88388
#> 483  1.964965 -20.81781 24.88354
#> 484  1.966967 -20.81336 24.88333
#> 485  1.968969 -20.80904 24.88326
#> 486  1.970971 -20.80486 24.88332
#> 487  1.972973 -20.80082 24.88352
#> 488  1.974975 -20.79692 24.88386
#> 489  1.976977 -20.79315 24.88434
#> 490  1.978979 -20.78952 24.88495
#> 491  1.980981 -20.78602 24.88570
#> 492  1.982983 -20.78266 24.88659
#> 493  1.984985 -20.77944 24.88761
#> 494  1.986987 -20.77636 24.88877
#> 495  1.988989 -20.77342 24.89007
#> 496  1.990991 -20.77061 24.89151
#> 497  1.992993 -20.76794 24.89308
#> 498  1.994995 -20.76540 24.89479
#> 499  1.996997 -20.76301 24.89664
#> 500  1.998999 -20.76075 24.89862
#> 501  2.001001 -20.75862 24.90075
#> 502  2.003003 -20.75664 24.90301
#> 503  2.005005 -20.75479 24.90540
#> 504  2.007007 -20.75308 24.90794
#> 505  2.009009 -20.75151 24.91061
#> 506  2.011011 -20.75007 24.91342
#> 507  2.013013 -20.74877 24.91636
#> 508  2.015015 -20.74761 24.91944
#> 509  2.017017 -20.74659 24.92266
#> 510  2.019019 -20.74570 24.92602
#> 511  2.021021 -20.74495 24.92952
#> 512  2.023023 -20.74434 24.93315
#> 513  2.025025 -20.74386 24.93692
#> 514  2.027027 -20.74352 24.94082
#> 515  2.029029 -20.74332 24.94486
#> 516  2.031031 -20.74326 24.94904
#> 517  2.033033 -20.74333 24.95336
#> 518  2.035035 -20.74354 24.95781
#> 519  2.037037 -20.74388 24.96240
#> 520  2.039039 -20.74437 24.96713
#> 521  2.041041 -20.74499 24.97199
#> 522  2.043043 -20.74574 24.97699
#> 523  2.045045 -20.74664 24.98213
#> 524  2.047047 -20.74767 24.98741
#> 525  2.049049 -20.74883 24.99282
#> 526  2.051051 -20.75013 24.99836
#> 527  2.053053 -20.75157 25.00405
#> 528  2.055055 -20.75315 25.00987
#> 529  2.057057 -20.75486 25.01582
#> 530  2.059059 -20.75671 25.02191
#> 531  2.061061 -20.75869 25.02814
#> 532  2.063063 -20.76081 25.03451
#> 533  2.065065 -20.76307 25.04101
#> 534  2.067067 -20.76546 25.04765
#> 535  2.069069 -20.76799 25.05442
#> 536  2.071071 -20.77066 25.06133
#> 537  2.073073 -20.77346 25.06837
#> 538  2.075075 -20.77639 25.07555
#> 539  2.077077 -20.77946 25.08287
#> 540  2.079079 -20.78267 25.09032
#> 541  2.081081 -20.78601 25.09790
#> 542  2.083083 -20.78949 25.10562
#> 543  2.085085 -20.79310 25.11348
#> 544  2.087087 -20.79685 25.12147
#> 545  2.089089 -20.80073 25.12960
#> 546  2.091091 -20.80475 25.13786
#> 547  2.093093 -20.80890 25.14625
#> 548  2.095095 -20.81318 25.15479
#> 549  2.097097 -20.81760 25.16345
#> 550  2.099099 -20.82216 25.17225
#> 551  2.101101 -20.82685 25.18118
#> 552  2.103103 -20.83167 25.19025
#> 553  2.105105 -20.83663 25.19945
#> 554  2.107107 -20.84172 25.20879
#> 555  2.109109 -20.84695 25.21826
#> 556  2.111111 -20.85231 25.22786
#> 557  2.113113 -20.85780 25.23760
#> 558  2.115115 -20.86343 25.24747
#> 559  2.117117 -20.86918 25.25747
#> 560  2.119119 -20.87508 25.26761
#> 561  2.121121 -20.88110 25.27788
#> 562  2.123123 -20.88726 25.28828
#> 563  2.125125 -20.89355 25.29882
#> 564  2.127127 -20.89997 25.30948
#> 565  2.129129 -20.90653 25.32028
#> 566  2.131131 -20.91322 25.33122
#> 567  2.133133 -20.92004 25.34228
#> 568  2.135135 -20.92699 25.35348
#> 569  2.137137 -20.93407 25.36481
#> 570  2.139139 -20.94129 25.37627
#> 571  2.141141 -20.94864 25.38786
#> 572  2.143143 -20.95612 25.39958
#> 573  2.145145 -20.96373 25.41143
#> 574  2.147147 -20.97147 25.42342
#> 575  2.149149 -20.97934 25.43553
#> 576  2.151151 -20.98734 25.44778
#> 577  2.153153 -20.99547 25.46016
#> 578  2.155155 -21.00373 25.47266
#> 579  2.157157 -21.01213 25.48530
#> 580  2.159159 -21.02065 25.49807
#> 581  2.161161 -21.02930 25.51096
#> 582  2.163163 -21.03809 25.52399
#> 583  2.165165 -21.04700 25.53715
#> 584  2.167167 -21.05604 25.55043
#> 585  2.169169 -21.06521 25.56385
#> 586  2.171171 -21.07451 25.57739
#> 587  2.173173 -21.08393 25.59106
#> 588  2.175175 -21.09349 25.60486
#> 589  2.177177 -21.10318 25.61879
#> 590  2.179179 -21.11299 25.63285
#> 591  2.181181 -21.12293 25.64703
#> 592  2.183183 -21.13300 25.66134
#> 593  2.185185 -21.14319 25.67578
#> 594  2.187187 -21.15351 25.69035
#> 595  2.189189 -21.16396 25.70505
#> 596  2.191191 -21.17454 25.71987
#> 597  2.193193 -21.18524 25.73481
#> 598  2.195195 -21.19608 25.74989
#> 599  2.197197 -21.20703 25.76509
#> 600  2.199199 -21.21811 25.78042
#> 601  2.201201 -21.22932 25.79587
#> 602  2.203203 -21.24066 25.81145
#> 603  2.205205 -21.25212 25.82715
#> 604  2.207207 -21.26370 25.84298
#> 605  2.209209 -21.27541 25.85893
#> 606  2.211211 -21.28724 25.87501
#> 607  2.213213 -21.29920 25.89121
#> 608  2.215215 -21.31129 25.90754
#> 609  2.217217 -21.32349 25.92399
#> 610  2.219219 -21.33582 25.94057
#> 611  2.221221 -21.34828 25.95727
#> 612  2.223223 -21.36086 25.97409
#> 613  2.225225 -21.37356 25.99104
#> 614  2.227227 -21.38638 26.00810
#> 615  2.229229 -21.39933 26.02530
#> 616  2.231231 -21.41240 26.04261
#> 617  2.233233 -21.42559 26.06004
#> 618  2.235235 -21.43890 26.07760
#> 619  2.237237 -21.45234 26.09528
#> 620  2.239239 -21.46590 26.11308
#> 621  2.241241 -21.47958 26.13101
#> 622  2.243243 -21.49337 26.14905
#> 623  2.245245 -21.50730 26.16721
#> 624  2.247247 -21.52134 26.18550
#> 625  2.249249 -21.53550 26.20391
#> 626  2.251251 -21.54978 26.22243
#> 627  2.253253 -21.56418 26.24108
#> 628  2.255255 -21.57870 26.25984
#> 629  2.257257 -21.59334 26.27873
#> 630  2.259259 -21.60810 26.29773
#> 631  2.261261 -21.62298 26.31685
#> 632  2.263263 -21.63797 26.33609
#> 633  2.265265 -21.65309 26.35545
#> 634  2.267267 -21.66832 26.37493
#> 635  2.269269 -21.68367 26.39452
#> 636  2.271271 -21.69914 26.41424
#> 637  2.273273 -21.71473 26.43407
#> 638  2.275275 -21.73043 26.45401
#> 639  2.277277 -21.74625 26.47408
#> 640  2.279279 -21.76218 26.49426
#> 641  2.281281 -21.77824 26.51455
#> 642  2.283283 -21.79440 26.53497
#> 643  2.285285 -21.81069 26.55549
#> 644  2.287287 -21.82709 26.57614
#> 645  2.289289 -21.84360 26.59690
#> 646  2.291291 -21.86023 26.61777
#> 647  2.293293 -21.87698 26.63876
#> 648  2.295295 -21.89384 26.65986
#> 649  2.297297 -21.91081 26.68108
#> 650  2.299299 -21.92789 26.70241
#> 651  2.301301 -21.94509 26.72385
#> 652  2.303303 -21.96241 26.74541
#> 653  2.305305 -21.97984 26.76708
#> 654  2.307307 -21.99737 26.78887
#> 655  2.309309 -22.01503 26.81076
#> 656  2.311311 -22.03279 26.83277
#> 657  2.313313 -22.05067 26.85489
#> 658  2.315315 -22.06865 26.87712
#> 659  2.317317 -22.08675 26.89947
#> 660  2.319319 -22.10496 26.92192
#> 661  2.321321 -22.12329 26.94449
#> 662  2.323323 -22.14172 26.96716
#> 663  2.325325 -22.16026 26.98995
#> 664  2.327327 -22.17891 27.01285
#> 665  2.329329 -22.19767 27.03585
#> 666  2.331331 -22.21655 27.05897
#> 667  2.333333 -22.23553 27.08219
#> 668  2.335335 -22.25462 27.10553
#> 669  2.337337 -22.27381 27.12897
#> 670  2.339339 -22.29312 27.15252
#> 671  2.341341 -22.31254 27.17618
#> 672  2.343343 -22.33206 27.19995
#> 673  2.345345 -22.35169 27.22382
#> 674  2.347347 -22.37143 27.24780
#> 675  2.349349 -22.39127 27.27189
#> 676  2.351351 -22.41122 27.29609
#> 677  2.353353 -22.43128 27.32039
#> 678  2.355355 -22.45144 27.34479
#> 679  2.357357 -22.47171 27.36931
#> 680  2.359359 -22.49208 27.39393
#> 681  2.361361 -22.51256 27.41865
#> 682  2.363363 -22.53315 27.44348
#> 683  2.365365 -22.55384 27.46841
#> 684  2.367367 -22.57463 27.49345
#> 685  2.369369 -22.59553 27.51859
#> 686  2.371371 -22.61653 27.54384
#> 687  2.373373 -22.63764 27.56919
#> 688  2.375375 -22.65885 27.59464
#> 689  2.377377 -22.68016 27.62020
#> 690  2.379379 -22.70157 27.64586
#> 691  2.381381 -22.72309 27.67162
#> 692  2.383383 -22.74471 27.69748
#> 693  2.385385 -22.76643 27.72344
#> 694  2.387387 -22.78825 27.74951
#> 695  2.389389 -22.81017 27.77568
#> 696  2.391391 -22.83220 27.80195
#> 697  2.393393 -22.85432 27.82831
#> 698  2.395395 -22.87655 27.85478
#> 699  2.397397 -22.89887 27.88135
#> 700  2.399399 -22.92130 27.90802
#> 701  2.401401 -22.94382 27.93479
#> 702  2.403403 -22.96644 27.96166
#> 703  2.405405 -22.98917 27.98863
#> 704  2.407407 -23.01199 28.01569
#> 705  2.409409 -23.03491 28.04285
#> 706  2.411411 -23.05792 28.07012
#> 707  2.413413 -23.08104 28.09748
#> 708  2.415415 -23.10425 28.12493
#> 709  2.417417 -23.12756 28.15249
#> 710  2.419419 -23.15097 28.18014
#> 711  2.421421 -23.17447 28.20789
#> 712  2.423423 -23.19807 28.23573
#> 713  2.425425 -23.22177 28.26367
#> 714  2.427427 -23.24556 28.29170
#> 715  2.429429 -23.26945 28.31984
#> 716  2.431431 -23.29343 28.34806
#> 717  2.433433 -23.31750 28.37638
#> 718  2.435435 -23.34168 28.40480
#> 719  2.437437 -23.36594 28.43331
#> 720  2.439439 -23.39030 28.46191
#> 721  2.441441 -23.41475 28.49061
#> 722  2.443443 -23.43930 28.51940
#> 723  2.445445 -23.46394 28.54829
#> 724  2.447447 -23.48867 28.57726
#> 725  2.449449 -23.51350 28.60633
#> 726  2.451451 -23.53842 28.63549
#> 727  2.453453 -23.56343 28.66475
#> 728  2.455455 -23.58853 28.69409
#> 729  2.457457 -23.61372 28.72353
#> 730  2.459459 -23.63900 28.75306
#> 731  2.461461 -23.66438 28.78268
#> 732  2.463463 -23.68985 28.81239
#> 733  2.465465 -23.71540 28.84219
#> 734  2.467467 -23.74105 28.87208
#> 735  2.469469 -23.76678 28.90206
#> 736  2.471471 -23.79261 28.93213
#> 737  2.473473 -23.81852 28.96228
#> 738  2.475475 -23.84452 28.99253
#> 739  2.477477 -23.87061 29.02287
#> 740  2.479479 -23.89679 29.05329
#> 741  2.481481 -23.92306 29.08380
#> 742  2.483483 -23.94942 29.11440
#> 743  2.485485 -23.97586 29.14509
#> 744  2.487487 -24.00239 29.17586
#> 745  2.489489 -24.02901 29.20672
#> 746  2.491491 -24.05571 29.23767
#> 747  2.493493 -24.08250 29.26871
#> 748  2.495495 -24.10938 29.29983
#> 749  2.497497 -24.13634 29.33103
#> 750  2.499499 -24.16338 29.36232
#> 751  2.501502 -24.19052 29.39370
#> 752  2.503504 -24.21773 29.42516
#> 753  2.505506 -24.24503 29.45671
#> 754  2.507508 -24.27242 29.48834
#> 755  2.509510 -24.29989 29.52005
#> 756  2.511512 -24.32745 29.55185
#> 757  2.513514 -24.35508 29.58373
#> 758  2.515516 -24.38280 29.61570
#> 759  2.517518 -24.41061 29.64774
#> 760  2.519520 -24.43849 29.67987
#> 761  2.521522 -24.46646 29.71209
#> 762  2.523524 -24.49451 29.74438
#> 763  2.525526 -24.52265 29.77676
#> 764  2.527528 -24.55086 29.80922
#> 765  2.529530 -24.57916 29.84176
#> 766  2.531532 -24.60753 29.87438
#> 767  2.533534 -24.63599 29.90708
#> 768  2.535536 -24.66453 29.93986
#> 769  2.537538 -24.69315 29.97273
#> 770  2.539540 -24.72185 30.00567
#> 771  2.541542 -24.75062 30.03869
#> 772  2.543544 -24.77948 30.07179
#> 773  2.545546 -24.80842 30.10498
#> 774  2.547548 -24.83743 30.13824
#> 775  2.549550 -24.86653 30.17157
#> 776  2.551552 -24.89570 30.20499
#> 777  2.553554 -24.92495 30.23849
#> 778  2.555556 -24.95428 30.27206
#> 779  2.557558 -24.98369 30.30571
#> 780  2.559560 -25.01317 30.33944
#> 781  2.561562 -25.04273 30.37324
#> 782  2.563564 -25.07237 30.40712
#> 783  2.565566 -25.10208 30.44108
#> 784  2.567568 -25.13187 30.47512
#> 785  2.569570 -25.16174 30.50923
#> 786  2.571572 -25.19168 30.54341
#> 787  2.573574 -25.22170 30.57767
#> 788  2.575576 -25.25179 30.61201
#> 789  2.577578 -25.28196 30.64642
#> 790  2.579580 -25.31220 30.68091
#> 791  2.581582 -25.34252 30.71547
#> 792  2.583584 -25.37291 30.75010
#> 793  2.585586 -25.40337 30.78481
#> 794  2.587588 -25.43391 30.81959
#> 795  2.589590 -25.46452 30.85445
#> 796  2.591592 -25.49520 30.88938
#> 797  2.593594 -25.52596 30.92438
#> 798  2.595596 -25.55679 30.95945
#> 799  2.597598 -25.58769 30.99460
#> 800  2.599600 -25.61867 31.02982
#> 801  2.601602 -25.64972 31.06511
#> 802  2.603604 -25.68083 31.10047
#> 803  2.605606 -25.71202 31.13591
#> 804  2.607608 -25.74328 31.17141
#> 805  2.609610 -25.77462 31.20699
#> 806  2.611612 -25.80602 31.24263
#> 807  2.613614 -25.83749 31.27835
#> 808  2.615616 -25.86903 31.31414
#> 809  2.617618 -25.90065 31.35000
#> 810  2.619620 -25.93233 31.38592
#> 811  2.621622 -25.96408 31.42192
#> 812  2.623624 -25.99590 31.45798
#> 813  2.625626 -26.02779 31.49412
#> 814  2.627628 -26.05975 31.53032
#> 815  2.629630 -26.09177 31.56659
#> 816  2.631632 -26.12387 31.60293
#> 817  2.633634 -26.15603 31.63933
#> 818  2.635636 -26.18826 31.67581
#> 819  2.637638 -26.22056 31.71235
#> 820  2.639640 -26.25292 31.74896
#> 821  2.641642 -26.28536 31.78564
#> 822  2.643644 -26.31786 31.82238
#> 823  2.645646 -26.35042 31.85919
#> 824  2.647648 -26.38305 31.89606
#> 825  2.649650 -26.41575 31.93301
#> 826  2.651652 -26.44851 31.97001
#> 827  2.653654 -26.48134 32.00709
#> 828  2.655656 -26.51423 32.04422
#> 829  2.657658 -26.54719 32.08143
#> 830  2.659660 -26.58022 32.11870
#> 831  2.661662 -26.61331 32.15603
#> 832  2.663664 -26.64646 32.19343
#> 833  2.665666 -26.67968 32.23089
#> 834  2.667668 -26.71296 32.26841
#> 835  2.669670 -26.74630 32.30600
#> 836  2.671672 -26.77971 32.34365
#> 837  2.673674 -26.81318 32.38137
#> 838  2.675676 -26.84671 32.41915
#> 839  2.677678 -26.88031 32.45699
#> 840  2.679680 -26.91397 32.49489
#> 841  2.681682 -26.94769 32.53286
#> 842  2.683684 -26.98148 32.57089
#> 843  2.685686 -27.01532 32.60898
#> 844  2.687688 -27.04923 32.64713
#> 845  2.689690 -27.08320 32.68534
#> 846  2.691692 -27.11723 32.72361
#> 847  2.693694 -27.15132 32.76195
#> 848  2.695696 -27.18547 32.80035
#> 849  2.697698 -27.21968 32.83880
#> 850  2.699700 -27.25396 32.87732
#> 851  2.701702 -27.28829 32.91590
#> 852  2.703704 -27.32268 32.95453
#> 853  2.705706 -27.35713 32.99323
#> 854  2.707708 -27.39164 33.03198
#> 855  2.709710 -27.42622 33.07080
#> 856  2.711712 -27.46084 33.10967
#> 857  2.713714 -27.49553 33.14861
#> 858  2.715716 -27.53028 33.18760
#> 859  2.717718 -27.56509 33.22665
#> 860  2.719720 -27.59995 33.26576
#> 861  2.721722 -27.63487 33.30492
#> 862  2.723724 -27.66985 33.34415
#> 863  2.725726 -27.70489 33.38343
#> 864  2.727728 -27.73998 33.42276
#> 865  2.729730 -27.77513 33.46216
#> 866  2.731732 -27.81034 33.50161
#> 867  2.733734 -27.84561 33.54112
#> 868  2.735736 -27.88093 33.58069
#> 869  2.737738 -27.91631 33.62031
#> 870  2.739740 -27.95174 33.65999
#> 871  2.741742 -27.98723 33.69972
#> 872  2.743744 -28.02277 33.73951
#> 873  2.745746 -28.05837 33.77936
#> 874  2.747748 -28.09403 33.81926
#> 875  2.749750 -28.12974 33.85921
#> 876  2.751752 -28.16551 33.89922
#> 877  2.753754 -28.20133 33.93929
#> 878  2.755756 -28.23720 33.97940
#> 879  2.757758 -28.27313 34.01958
#> 880  2.759760 -28.30912 34.05981
#> 881  2.761762 -28.34515 34.10009
#> 882  2.763764 -28.38124 34.14042
#> 883  2.765766 -28.41739 34.18081
#> 884  2.767768 -28.45358 34.22125
#> 885  2.769770 -28.48984 34.26175
#> 886  2.771772 -28.52614 34.30229
#> 887  2.773774 -28.56249 34.34289
#> 888  2.775776 -28.59890 34.38355
#> 889  2.777778 -28.63536 34.42425
#> 890  2.779780 -28.67188 34.46501
#> 891  2.781782 -28.70844 34.50582
#> 892  2.783784 -28.74506 34.54668
#> 893  2.785786 -28.78173 34.58759
#> 894  2.787788 -28.81845 34.62856
#> 895  2.789790 -28.85522 34.66957
#> 896  2.791792 -28.89204 34.71064
#> 897  2.793794 -28.92891 34.75175
#> 898  2.795796 -28.96583 34.79292
#> 899  2.797798 -29.00281 34.83414
#> 900  2.799800 -29.03983 34.87541
#> 901  2.801802 -29.07690 34.91672
#> 902  2.803804 -29.11403 34.95809
#> 903  2.805806 -29.15120 34.99951
#> 904  2.807808 -29.18842 35.04098
#> 905  2.809810 -29.22570 35.08249
#> 906  2.811812 -29.26302 35.12406
#> 907  2.813814 -29.30039 35.16567
#> 908  2.815816 -29.33780 35.20733
#> 909  2.817818 -29.37527 35.24905
#> 910  2.819820 -29.41279 35.29081
#> 911  2.821822 -29.45035 35.33261
#> 912  2.823824 -29.48796 35.37447
#> 913  2.825826 -29.52562 35.41637
#> 914  2.827828 -29.56333 35.45833
#> 915  2.829830 -29.60109 35.50033
#> 916  2.831832 -29.63889 35.54237
#> 917  2.833834 -29.67674 35.58447
#> 918  2.835836 -29.71464 35.62661
#> 919  2.837838 -29.75258 35.66880
#> 920  2.839840 -29.79057 35.71103
#> 921  2.841842 -29.82861 35.75331
#> 922  2.843844 -29.86669 35.79564
#> 923  2.845846 -29.90482 35.83801
#> 924  2.847848 -29.94299 35.88043
#> 925  2.849850 -29.98121 35.92290
#> 926  2.851852 -30.01948 35.96541
#> 927  2.853854 -30.05779 36.00796
#> 928  2.855856 -30.09615 36.05056
#> 929  2.857858 -30.13455 36.09321
#> 930  2.859860 -30.17300 36.13590
#> 931  2.861862 -30.21149 36.17864
#> 932  2.863864 -30.25003 36.22142
#> 933  2.865866 -30.28861 36.26425
#> 934  2.867868 -30.32724 36.30712
#> 935  2.869870 -30.36591 36.35003
#> 936  2.871872 -30.40462 36.39299
#> 937  2.873874 -30.44338 36.43599
#> 938  2.875876 -30.48218 36.47904
#> 939  2.877878 -30.52102 36.52213
#> 940  2.879880 -30.55991 36.56526
#> 941  2.881882 -30.59884 36.60843
#> 942  2.883884 -30.63782 36.65165
#> 943  2.885886 -30.67684 36.69491
#> 944  2.887888 -30.71590 36.73822
#> 945  2.889890 -30.75500 36.78157
#> 946  2.891892 -30.79414 36.82495
#> 947  2.893894 -30.83333 36.86839
#> 948  2.895896 -30.87256 36.91186
#> 949  2.897898 -30.91183 36.95538
#> 950  2.899900 -30.95115 36.99893
#> 951  2.901902 -30.99050 37.04253
#> 952  2.903904 -31.02990 37.08617
#> 953  2.905906 -31.06933 37.12985
#> 954  2.907908 -31.10881 37.17358
#> 955  2.909910 -31.14833 37.21734
#> 956  2.911912 -31.18789 37.26115
#> 957  2.913914 -31.22750 37.30499
#> 958  2.915916 -31.26714 37.34888
#> 959  2.917918 -31.30682 37.39281
#> 960  2.919920 -31.34654 37.43677
#> 961  2.921922 -31.38631 37.48078
#> 962  2.923924 -31.42611 37.52483
#> 963  2.925926 -31.46595 37.56891
#> 964  2.927928 -31.50583 37.61304
#> 965  2.929930 -31.54576 37.65721
#> 966  2.931932 -31.58572 37.70141
#> 967  2.933934 -31.62572 37.74566
#> 968  2.935936 -31.66576 37.78994
#> 969  2.937938 -31.70584 37.83427
#> 970  2.939940 -31.74596 37.87863
#> 971  2.941942 -31.78611 37.92303
#> 972  2.943944 -31.82631 37.96747
#> 973  2.945946 -31.86654 38.01195
#> 974  2.947948 -31.90681 38.05646
#> 975  2.949950 -31.94712 38.10102
#> 976  2.951952 -31.98747 38.14561
#> 977  2.953954 -32.02786 38.19024
#> 978  2.955956 -32.06828 38.23491
#> 979  2.957958 -32.10874 38.27961
#> 980  2.959960 -32.14924 38.32435
#> 981  2.961962 -32.18978 38.36914
#> 982  2.963964 -32.23035 38.41395
#> 983  2.965966 -32.27096 38.45881
#> 984  2.967968 -32.31161 38.50370
#> 985  2.969970 -32.35229 38.54863
#> 986  2.971972 -32.39301 38.59359
#> 987  2.973974 -32.43377 38.63859
#> 988  2.975976 -32.47456 38.68363
#> 989  2.977978 -32.51539 38.72870
#> 990  2.979980 -32.55626 38.77381
#> 991  2.981982 -32.59716 38.81896
#> 992  2.983984 -32.63810 38.86414
#> 993  2.985986 -32.67907 38.90936
#> 994  2.987988 -32.72008 38.95461
#> 995  2.989990 -32.76112 38.99990
#> 996  2.991992 -32.80220 39.04523
#> 997  2.993994 -32.84332 39.09059
#> 998  2.995996 -32.88447 39.13598
#> 999  2.997998 -32.92565 39.18141
#> 1000 3.000000 -32.96687 39.22687